Add support for analogue input to Switch library
Due to the limited GPIO on the EPS8266, it is necessary to use the
analogue input to read the state of the pushbutton on the rotary
encoder. The Switch library was hard-coded to use the digitalRead()
function - this commit adds a compile-time option to use analogRead()
and do a threshold comparison instead.