Implement complete lamp functionality in example
Drawing on older code which used ESP-link, this commit adds full lamp
output control using the rotary encoder - this means turning the
encoder changes the dimming level, and pressing the encoder switch turns
the output on or off. Delayed off timer is also implemented, as is the
mode setting function (although this does not have any affect at the
moment). No wireless functions are implemented yet, but from this commit
onwards a WiLED lamp can be used with proper function.