Begin creating a hardware abstraction layer
This commit begins to introduce the notion of a hardware abstraction
layer - initially with an abstraction for the LED output. Instead of
directly using 'analogWrite' in the LEDOutput class, for example, there
is a generic function 'setPWM' provided by the HAL that then calls the
appropriate 'analogWrite' call for the specific hardware. This doesn't
add much for just the ESP8266 - but it will make isolating Arduino
specific code much easier.